Enable LinearSeqOptimizedTest
